1.rename ConStants.h to DrConstants.h 2. fix bug to support negative number  for yAxis 3. give more interface to set the chart
