+++ dr-charts/Classes/Legend/LegendView.m
@@ -7,7 +7,7 @@
 //
 
 #import "LegendView.h"
-#import "Constants.h"
+#import "DrConstants.h"
 
 @implementation LegendView
